+++ CHANGELOG.md
@@ -2,5 +2,6 @@ Changelog
 =========
  
 ## 1.0.0 (2019-12-30)
+ * Added debug in validator.
  * Bugfix client validation for inivisible captcha.
  * First release. Added widget and validator.